Abort client operation mark deadlock bugfix
Imagine a server that comes into leadership, but hasnt commmited the
last the operation of a client accepts the operation from the same
client. If the server would step down from leadership before the
previously committed command is committed by this server, the latest
command would get marked as the most recent command performed by this
server for the client.

We'll only track completed commands for most recent commands as aborted
commands will get retried again and again until completed